common interface file for reshuffling
